refactor(spanner): reorganize public API into logical domain modules
Reorganizes the public API of the Spanner client library to replace
direct crate-root type re-exports with a domain-oriented public module
structure.

Core types are consolidated into logical public namespaces:
- `key`: Key and key range types (e.g., `Key`, `KeySet`, `KeyRange`).
- `mutation`: Mutation builders (e.g., `Mutation`).
- `read`: Read requests (e.g., `ReadRequest`).
- `result`: Result streams and rows (e.g., `ResultSet`, `Row`).
- `statement`: SQL statement and parameter binding types (e.g., `Statement`).
- `value`: Data types and conversions (e.g., `Value`, `ToValue`, `FromValue`).

This logical domain split improves API discoverability, avoids API surface
bloat at the crate root, and aligns with the modular design of other
handwritten crates like Pub/Sub.
