+++ src/spanner/examples/src/read/read_data.rs
@@ -14,7 +14,8 @@
 
 // [START spanner_read_data]
 use google_cloud_spanner::client::DatabaseClient;
-use google_cloud_spanner::{KeySet, ReadRequest};
+use google_cloud_spanner::key::KeySet;
+use google_cloud_spanner::read::ReadRequest;
 
 pub async fn sample(client: &DatabaseClient) -> anyhow::Result<()> {
     // Build a read request to read all rows in the Albums table
