+++ src/spanner/src/partitioned_dml_transaction.rs
@@ -31,7 +31,7 @@ use google_cloud_gax::options::RequestOptions as GaxRequestOptions;
 /// # Example
 /// ```
 /// # use google_cloud_spanner::client::Spanner;
-/// # use google_cloud_spanner::Statement;
+/// # use google_cloud_spanner::statement::Statement;
 /// # async fn build_transaction(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
 ///     let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
 ///     let transaction = db_client.partitioned_dml_transaction().build().await?;
@@ -145,7 +145,7 @@ impl PartitionedDmlTransaction {
     /// # Example
     /// ```
     /// # use google_cloud_spanner::client::Spanner;
-    /// # use google_cloud_spanner::Statement;
+    /// # use google_cloud_spanner::statement::Statement;
     /// # async fn run(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let transaction = db_client.partitioned_dml_transaction().build().await?;
