+++ src/spanner/src/read.rs
@@ -25,8 +25,8 @@ use std::time::Duration;
 ///
 /// # Example
 /// ```
-/// # use google_cloud_spanner::ReadRequest;
-/// # use google_cloud_spanner::KeySet;
+/// # use google_cloud_spanner::read::ReadRequest;
+/// # use google_cloud_spanner::key::KeySet;
 /// # use google_cloud_spanner::key;
 /// // Read all rows from a table using its primary key.
 /// let read_all = ReadRequest::builder("Users", vec!["Id", "Name"])
@@ -54,7 +54,8 @@ impl ReadRequestBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::{ReadRequest, KeySet};
+    /// # use google_cloud_spanner::read::ReadRequest;
+    /// # use google_cloud_spanner::key::KeySet;
     /// let request = ReadRequest::builder("Users", vec!["Id", "Name"]).with_keys(KeySet::all());
     /// ```
     ///
@@ -79,7 +80,8 @@ impl ReadRequestBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::{ReadRequest, KeySet};
+    /// # use google_cloud_spanner::read::ReadRequest;
+    /// # use google_cloud_spanner::key::KeySet;
     /// # use google_cloud_spanner::key;
     /// let request = ReadRequest::builder("Users", vec!["Id", "Name"]).with_index("UsersByIndex", key![1_i64]);
     /// ```
@@ -126,7 +128,8 @@ impl ConfiguredReadRequestBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::{ReadRequest, KeySet};
+    /// # use google_cloud_spanner::read::ReadRequest;
+    /// # use google_cloud_spanner::key::KeySet;
     /// let request = ReadRequest::builder("Users", vec!["Id"])
     ///     .with_keys(KeySet::all())
     ///     .set_limit(10)
@@ -143,7 +146,8 @@ impl ConfiguredReadRequestBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::{ReadRequest, KeySet};
+    /// # use google_cloud_spanner::read::ReadRequest;
+    /// # use google_cloud_spanner::key::KeySet;
     /// let request = ReadRequest::builder("Users", vec!["Id"])
     ///     .with_keys(KeySet::all())
     ///     .set_request_tag("my-tag")
@@ -162,7 +166,8 @@ impl ConfiguredReadRequestBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::{ReadRequest, KeySet};
+    /// # use google_cloud_spanner::read::ReadRequest;
+    /// # use google_cloud_spanner::key::KeySet;
     /// # use google_cloud_spanner::model::request_options::Priority;
     /// let request = ReadRequest::builder("Users", vec!["Id"])
     ///     .with_keys(KeySet::all())
@@ -179,8 +184,8 @@ impl ConfiguredReadRequestBuilder {
     /// Sets the directed read options for this request.
     ///
     /// ```
-    /// # use google_cloud_spanner::ReadRequest;
-    /// # use google_cloud_spanner::KeySet;
+    /// # use google_cloud_spanner::read::ReadRequest;
+    /// # use google_cloud_spanner::key::KeySet;
     /// # use google_cloud_spanner::model::DirectedReadOptions;
     /// let dro = DirectedReadOptions::default();
     /// let req = ReadRequest::builder("MyTable", vec!["col1"])
@@ -200,7 +205,8 @@ impl ConfiguredReadRequestBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::{ReadRequest, KeySet};
+    /// # use google_cloud_spanner::read::ReadRequest;
+    /// # use google_cloud_spanner::key::KeySet;
     /// # use google_cloud_spanner::model::read_request::OrderBy;
     /// let request = ReadRequest::builder("Users", vec!["Id"])
     ///     .with_keys(KeySet::all())
@@ -223,7 +229,8 @@ impl ConfiguredReadRequestBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::{ReadRequest, KeySet};
+    /// # use google_cloud_spanner::read::ReadRequest;
+    /// # use google_cloud_spanner::key::KeySet;
     /// # use google_cloud_spanner::model::read_request::LockHint;
     /// let request = ReadRequest::builder("Users", vec!["Id"])
     ///     .with_keys(KeySet::all())
@@ -305,7 +312,7 @@ impl ReadRequest {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::ReadRequest;
+    /// # use google_cloud_spanner::read::ReadRequest;
     /// let builder = ReadRequest::builder("Users", vec!["Id", "Name"]);
     /// ```
     ///
