+++ src/spanner/src/read_only_transaction.rs
@@ -35,7 +35,7 @@ use tokio::sync::Notify;
 /// # Example
 /// ```
 /// # use google_cloud_spanner::client::Spanner;
-/// # use google_cloud_spanner::TimestampBound;
+/// # use google_cloud_spanner::transaction::TimestampBound;
 /// # async fn build_tx(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
 /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
 /// let read_only_tx = db_client.single_use()
@@ -62,7 +62,7 @@ impl SingleUseReadOnlyTransactionBuilder {
     /// # Example
     /// ```
     /// # use google_cloud_spanner::client::Spanner;
-    /// # use google_cloud_spanner::TimestampBound;
+    /// # use google_cloud_spanner::transaction::TimestampBound;
     /// # async fn set_bound(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let builder = db_client.single_use().set_timestamp_bound(TimestampBound::strong());
@@ -123,7 +123,7 @@ impl SingleUseReadOnlyTransactionBuilder {
 /// # Example
 /// ```
 /// # use google_cloud_spanner::client::Spanner;
-/// # use google_cloud_spanner::Statement;
+/// # use google_cloud_spanner::statement::Statement;
 /// # async fn run(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
 /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
 /// let tx = db_client.single_use().build();
@@ -145,7 +145,7 @@ impl SingleUseReadOnlyTransaction {
     /// # Example
     /// ```
     /// # use google_cloud_spanner::client::Spanner;
-    /// # use google_cloud_spanner::Statement;
+    /// # use google_cloud_spanner::statement::Statement;
     /// # async fn run(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let tx = db_client.single_use().build();
@@ -172,7 +172,8 @@ impl SingleUseReadOnlyTransaction {
     /// # Example
     /// ```
     /// # use google_cloud_spanner::client::Spanner;
-    /// # use google_cloud_spanner::{KeySet, ReadRequest};
+    /// # use google_cloud_spanner::key::KeySet;
+    /// # use google_cloud_spanner::read::ReadRequest;
     /// # use google_cloud_spanner::key;
     /// # async fn run(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -222,7 +223,7 @@ pub enum BeginTransactionOption {
 /// # Example
 /// ```
 /// # use google_cloud_spanner::client::Spanner;
-/// # use google_cloud_spanner::TimestampBound;
+/// # use google_cloud_spanner::transaction::TimestampBound;
 /// # async fn build_tx(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
 /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
 /// let read_only_tx = db_client.read_only_transaction()
@@ -255,7 +256,7 @@ impl MultiUseReadOnlyTransactionBuilder {
     /// ```
     /// # use google_cloud_spanner::client::Spanner;
     /// # use google_cloud_spanner::transaction::BeginTransactionOption;
-    /// # use google_cloud_spanner::Statement;
+    /// # use google_cloud_spanner::statement::Statement;
     /// # async fn set_begin_option(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let transaction = db_client.read_only_transaction().with_begin_transaction_option(BeginTransactionOption::ExplicitBegin).build().await?;
@@ -362,7 +363,7 @@ impl MultiUseReadOnlyTransactionBuilder {
     /// # Example
     /// ```
     /// # use google_cloud_spanner::client::Spanner;
-    /// # use google_cloud_spanner::TimestampBound;
+    /// # use google_cloud_spanner::transaction::TimestampBound;
     /// # async fn set_bound(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let builder = db_client.read_only_transaction().set_timestamp_bound(TimestampBound::strong());
@@ -457,7 +458,7 @@ impl MultiUseReadOnlyTransactionBuilder {
 /// # Example
 /// ```
 /// # use google_cloud_spanner::client::Spanner;
-/// # use google_cloud_spanner::Statement;
+/// # use google_cloud_spanner::statement::Statement;
 /// # async fn run(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
 /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
 /// let tx = db_client.read_only_transaction().build().await?;
@@ -488,7 +489,7 @@ impl MultiUseReadOnlyTransaction {
     /// # Example
     /// ```
     /// # use google_cloud_spanner::client::Spanner;
-    /// # use google_cloud_spanner::Statement;
+    /// # use google_cloud_spanner::statement::Statement;
     /// # async fn run(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let tx = db_client.read_only_transaction().build().await?;
@@ -515,7 +516,8 @@ impl MultiUseReadOnlyTransaction {
     /// # Example
     /// ```
     /// # use google_cloud_spanner::client::Spanner;
-    /// # use google_cloud_spanner::{KeySet, ReadRequest};
+    /// # use google_cloud_spanner::key::KeySet;
+    /// # use google_cloud_spanner::read::ReadRequest;
     /// # use google_cloud_spanner::key;
     /// # async fn run(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -1101,9 +1103,9 @@ impl ReadContext {
 #[cfg(test)]
 pub(crate) mod tests {
     use super::*;
-    use crate::Statement;
     use crate::result_set::tests::adapt;
     use crate::result_set::tests::string_val;
+    use crate::statement::Statement;
     use crate::value::Value;
     use gaxi::grpc::tonic::{self, Code, Response, Status};
     use google_cloud_gax::error::rpc::Code as GaxCode;
@@ -1233,7 +1235,7 @@ pub(crate) mod tests {
     #[tokio_test_no_panics]
     async fn execute_single_query() {
         use super::super::result_set::tests::string_val;
-        use crate::Statement;
+        use crate::statement::Statement;
         use crate::value::Value;
 
         let mut mock = create_session_mock();
@@ -1268,7 +1270,7 @@ pub(crate) mod tests {
     #[tokio_test_no_panics]
     async fn execute_multi_query() {
         use super::super::result_set::tests::string_val;
-        use crate::Statement;
+        use crate::statement::Statement;
         use crate::value::Value;
         use spanner_grpc_mock::google::spanner::v1 as mock_v1;
 
@@ -1344,7 +1346,7 @@ pub(crate) mod tests {
     #[tokio_test_no_panics]
     async fn execute_multi_query_inline_begin() -> anyhow::Result<()> {
         use super::super::result_set::tests::string_val;
-        use crate::Statement;
+        use crate::statement::Statement;
         use crate::value::Value;
         use spanner_grpc_mock::google::spanner::v1 as mock_v1;
 
@@ -1438,8 +1440,9 @@ pub(crate) mod tests {
     #[tokio_test_no_panics]
     async fn execute_single_read() {
         use super::super::result_set::tests::string_val;
+        use crate::key::KeySet;
+        use crate::read::ReadRequest;
         use crate::value::Value;
-        use crate::{KeySet, ReadRequest};
 
         let mut mock = create_session_mock();
 
@@ -1474,8 +1477,9 @@ pub(crate) mod tests {
     #[tokio_test_no_panics]
     async fn execute_multi_read() -> anyhow::Result<()> {
         use super::super::result_set::tests::string_val;
+        use crate::key::KeySet;
+        use crate::read::ReadRequest;
         use crate::value::Value;
-        use crate::{KeySet, ReadRequest};
         use spanner_grpc_mock::google::spanner::v1 as mock_v1;
 
         let mut mock = create_session_mock();
@@ -1750,8 +1754,9 @@ pub(crate) mod tests {
 
     #[tokio_test_no_panics]
     async fn inline_begin_read_failure_retry_success() -> anyhow::Result<()> {
+        use crate::key::KeySet;
+        use crate::read::ReadRequest;
         use crate::value::Value;
-        use crate::{KeySet, ReadRequest};
         use gaxi::grpc::tonic::Status;
         use tonic::Response;
 
@@ -1821,7 +1826,7 @@ pub(crate) mod tests {
 
     #[tokio_test_no_panics]
     async fn single_use_query_send_error_returns_immediately() -> anyhow::Result<()> {
-        use crate::Statement;
+        use crate::statement::Statement;
         use gaxi::grpc::tonic::Status;
 
         let mut mock = create_session_mock();
@@ -1850,7 +1855,7 @@ pub(crate) mod tests {
     #[tokio_test_no_panics]
     async fn inline_begin_already_started_query_send_error_returns_immediately()
     -> anyhow::Result<()> {
-        use crate::Statement;
+        use crate::statement::Statement;
         use gaxi::grpc::tonic::Status;
         use spanner_grpc_mock::google::spanner::v1 as mock_v1;
 
@@ -2414,7 +2419,8 @@ pub(crate) mod tests {
 
     #[tokio_test_no_panics]
     async fn execute_concurrent_reads_inline_begin() -> anyhow::Result<()> {
-        use crate::{KeySet, ReadRequest};
+        use crate::key::KeySet;
+        use crate::read::ReadRequest;
         let mut mock = create_session_mock();
         mock.expect_begin_transaction().never();
