+++ src/spanner/src/result_set_metadata.rs
@@ -14,12 +14,14 @@
 
 use std::sync::Arc;
 
-/// Metadata about a [`ResultSet`](crate::ResultSet).
+/// Metadata about a [`ResultSet`](crate::result::ResultSet).
 ///
 /// # Example
 ///
 /// ```
-/// # use google_cloud_spanner::{Spanner, Statement, TypeCode};
+/// # use google_cloud_spanner::client::Spanner;
+/// # use google_cloud_spanner::statement::Statement;
+/// # use google_cloud_spanner::value::TypeCode;
 /// # async fn test_doc() -> Result<(), Box<dyn std::error::Error>> {
 /// let client = Spanner::builder().build().await?;
 /// let db = client.database_client("projects/p/instances/i/databases/d").build().await?;
