+++ src/spanner/src/transaction_runner.rs
@@ -33,7 +33,8 @@ use wkt::Duration;
 ///
 /// # Example
 /// ```
-/// # use google_cloud_spanner::{Spanner, Statement};
+/// # use google_cloud_spanner::client::Spanner;
+/// # use google_cloud_spanner::statement::Statement;
 /// # async fn run(client: Spanner) -> Result<(), google_cloud_spanner::Error> {
 /// let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
 /// let runner = db_client.read_write_transaction().build().await?;
@@ -72,7 +73,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use std::time::Duration;
     /// # async fn run(client: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// # let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -97,7 +98,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use std::time::Duration;
     /// # async fn sample(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -121,7 +122,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use google_cloud_gax::retry_policy::NeverRetry;
     /// # async fn sample(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -145,7 +146,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use google_cloud_gax::exponential_backoff::ExponentialBackoff;
     /// # async fn sample(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -169,7 +170,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use std::time::Duration;
     /// # async fn sample(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -191,7 +192,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use google_cloud_gax::retry_policy::NeverRetry;
     /// # async fn sample(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -213,7 +214,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use google_cloud_gax::exponential_backoff::ExponentialBackoff;
     /// # async fn sample(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -235,7 +236,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use google_cloud_spanner::model::transaction_options::IsolationLevel;
     /// # async fn run(client: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -258,7 +259,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use google_cloud_spanner::model::transaction_options::read_write::ReadLockMode;
     /// # async fn run(client: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -281,7 +282,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # async fn build_tx(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let runner = db_client.read_write_transaction()
@@ -304,7 +305,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use google_cloud_spanner::transaction::BeginTransactionOption;
     /// # async fn run(client: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -341,7 +342,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use google_cloud_spanner::model::request_options::Priority;
     /// # async fn run(client: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -362,7 +363,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use wkt::Duration;
     /// # async fn run(client: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -389,7 +390,7 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # async fn build_tx(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let runner = db_client.read_write_transaction()
@@ -416,7 +417,8 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::{Spanner, Statement};
+    /// # use google_cloud_spanner::client::Spanner;
+    /// # use google_cloud_spanner::statement::Statement;
     /// # async fn run_tx(client: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// # let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let runner = db_client.read_write_transaction()
@@ -448,7 +450,7 @@ impl TransactionRunnerBuilder {
     /// # Example
     /// ```
     /// # use std::time::Duration;
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use google_cloud_spanner::transaction::BasicTransactionRetryPolicy;
     /// # async fn run(client: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -474,7 +476,8 @@ impl TransactionRunnerBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::{Spanner, Statement};
+    /// # use google_cloud_spanner::client::Spanner;
+    /// # use google_cloud_spanner::statement::Statement;
     /// # async fn run(client: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let runner = db_client.read_write_transaction().build().await?;
@@ -521,7 +524,8 @@ impl TransactionRunner {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::{Spanner, Statement};
+    /// # use google_cloud_spanner::client::Spanner;
+    /// # use google_cloud_spanner::statement::Statement;
     /// # async fn run_tx(client: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let runner = db_client.read_write_transaction().build().await?;
@@ -619,7 +623,7 @@ impl TransactionRunner {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::Mutation;
+    use crate::mutation::Mutation;
     use crate::read_only_transaction::tests::{create_session_mock, setup_db_client};
     use crate::transaction_retry_policy::tests::create_aborted_status;
     use gaxi::grpc::tonic;
