+++ src/spanner/src/write_only_transaction.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::Mutation;
 use crate::client::{DatabaseClient, amend_request_options_for_lar};
 use crate::model::request_options::Priority;
 use crate::model::transaction_options::ReadWrite;
@@ -20,6 +19,7 @@ use crate::model::{
     BeginTransactionRequest, CommitRequest, CommitResponse, MultiplexedSessionPrecommitToken,
     Mutation as ProtoMutation, RequestOptions, TransactionOptions,
 };
+use crate::mutation::Mutation;
 use crate::transaction_retry_policy::{
     BasicTransactionRetryPolicy, TransactionRetryPolicy, retry_aborted,
 };
@@ -101,7 +101,7 @@ impl WriteOnlyTransactionBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use wkt::Duration;
     /// # async fn sample(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -126,7 +126,7 @@ impl WriteOnlyTransactionBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # async fn build_tx(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let transaction = db_client.write_only_transaction()
@@ -152,7 +152,8 @@ impl WriteOnlyTransactionBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::{Mutation, Spanner};
+    /// # use google_cloud_spanner::mutation::Mutation;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # async fn test_doc() -> Result<(), Box<dyn std::error::Error>> {
     /// # let client = Spanner::builder().build().await?;
     /// # let db = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -184,7 +185,7 @@ impl WriteOnlyTransactionBuilder {
     /// # Example
     /// ```
     /// # use std::time::Duration;
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use google_cloud_spanner::transaction::BasicTransactionRetryPolicy;
     /// # async fn build_tx(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -212,7 +213,7 @@ impl WriteOnlyTransactionBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use std::time::Duration;
     /// # async fn sample(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -231,7 +232,7 @@ impl WriteOnlyTransactionBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # use google_cloud_gax::retry_policy::NeverRetry;
     /// # async fn sample(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -371,7 +372,8 @@ impl WriteOnlyTransaction {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::{Mutation, Spanner};
+    /// # use google_cloud_spanner::mutation::Mutation;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # async fn test_doc() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = Spanner::builder().build().await?;
     /// let db = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -492,7 +494,8 @@ impl WriteOnlyTransaction {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::{Mutation, Spanner};
+    /// # use google_cloud_spanner::mutation::Mutation;
+    /// # use google_cloud_spanner::client::Spanner;
     /// # async fn test_doc() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = Spanner::builder().build().await?;
     /// let db = client.database_client("projects/p/instances/i/databases/d").build().await?;
