+++ tests/spanner/src/batch_read_only_transaction.rs
@@ -16,8 +16,11 @@ use crate::client::create_database_client;
 use google_cloud_spanner::batch::Partition;
 use google_cloud_spanner::client::DatabaseClient;
 use google_cloud_spanner::key;
+use google_cloud_spanner::key::KeySet;
 use google_cloud_spanner::model::PartitionOptions;
-use google_cloud_spanner::{KeySet, Mutation, ReadRequest, Statement};
+use google_cloud_spanner::mutation::Mutation;
+use google_cloud_spanner::read::ReadRequest;
+use google_cloud_spanner::statement::Statement;
 use google_cloud_test_utils::resource_names::LowercaseAlphanumeric;
 use serde_json;
