+++ tests/spanner/src/batch_write.rs
@@ -16,7 +16,8 @@ use anyhow::Result;
 use google_cloud_gax::error::rpc::Code;
 use google_cloud_spanner::client::DatabaseClient;
 use google_cloud_spanner::model::BatchWriteResponse;
-use google_cloud_spanner::{Mutation, MutationGroup, Statement};
+use google_cloud_spanner::mutation::{Mutation, MutationGroup};
+use google_cloud_spanner::statement::Statement;
 use google_cloud_test_utils::resource_names::LowercaseAlphanumeric;
 use std::time::Duration;
 use tokio::time::sleep;
