+++ tests/spanner/src/concurrent_inline_begin.rs
@@ -16,8 +16,9 @@ use crate::client::{get_database_id, get_emulator_host, provision_emulator, upda
 use crate::test_proxy::{InterceptionResult, PassThroughProxy};
 use futures::future::BoxFuture;
 use google_cloud_spanner::client::Spanner;
+use google_cloud_spanner::result::{ResultSet, Row};
 use google_cloud_spanner::transaction::BeginTransactionOption;
-use google_cloud_spanner::{ResultSet, Row, TimestampBound};
+use google_cloud_spanner::transaction::TimestampBound;
 use google_cloud_test_utils::resource_names::LowercaseAlphanumeric;
 use http::{Request, Response, StatusCode, Uri};
 use http_body::Frame;
