+++ tests/spanner/src/directed_read.rs
@@ -13,9 +13,10 @@
 // limitations under the License.
 
 use google_cloud_spanner::client::DatabaseClient;
+use google_cloud_spanner::key::KeySet;
 use google_cloud_spanner::model::DirectedReadOptions;
 use google_cloud_spanner::model::directed_read_options::{IncludeReplicas, Replicas};
-use google_cloud_spanner::{KeySet, ReadRequest};
+use google_cloud_spanner::read::ReadRequest;
 
 pub async fn read_only_with_directed_read(db_client: &DatabaseClient) -> anyhow::Result<()> {
     let mut dro = DirectedReadOptions::default();
