+++ tests/spanner/src/dml_returning.rs
@@ -14,7 +14,8 @@
 
 use anyhow::Result;
 use google_cloud_spanner::client::DatabaseClient;
-use google_cloud_spanner::{Mutation, Statement};
+use google_cloud_spanner::mutation::Mutation;
+use google_cloud_spanner::statement::Statement;
 use google_cloud_test_utils::resource_names::LowercaseAlphanumeric;
 
 pub async fn dml_then_return_execute_query(db_client: &DatabaseClient) -> Result<()> {
