+++ tests/spanner/src/read_only_transaction_options.rs
@@ -13,8 +13,10 @@
 // limitations under the License.
 
 use google_cloud_spanner::client::DatabaseClient;
+use google_cloud_spanner::mutation::Mutation;
+use google_cloud_spanner::statement::Statement;
 use google_cloud_spanner::transaction::BeginTransactionOption;
-use google_cloud_spanner::{Mutation, Statement, TimestampBound};
+use google_cloud_spanner::transaction::TimestampBound;
 use google_cloud_test_utils::resource_names::LowercaseAlphanumeric;
 use google_cloud_wkt::Timestamp as WktTimestamp;
 use std::time::Duration;
