+++ tests/spanner/src/read_write_transaction.rs
@@ -14,7 +14,9 @@
 
 use google_cloud_spanner::client::DatabaseClient;
 use google_cloud_spanner::key;
-use google_cloud_spanner::{Mutation, ReadRequest, Statement};
+use google_cloud_spanner::mutation::Mutation;
+use google_cloud_spanner::read::ReadRequest;
+use google_cloud_spanner::statement::Statement;
 use google_cloud_test_utils::resource_names::LowercaseAlphanumeric;
 
 pub async fn successful_read_write_transaction(db_client: &DatabaseClient) -> anyhow::Result<()> {
