+++ tests/spanner/src/write.rs
@@ -16,7 +16,11 @@ use base64::Engine;
 use base64::prelude::BASE64_STANDARD;
 use google_cloud_spanner::Decimal;
 use google_cloud_spanner::client::DatabaseClient;
-use google_cloud_spanner::{Kind, Mutation, Statement, Value};
+use google_cloud_spanner::key::KeySet;
+use google_cloud_spanner::mutation::Mutation;
+use google_cloud_spanner::read::ReadRequest;
+use google_cloud_spanner::statement::Statement;
+use google_cloud_spanner::value::{Kind, Value};
 use prost_types::value::Kind as ProtoKind;
 use prost_types::{ListValue, Value as ProtoValue};
 use std::str::FromStr;
@@ -682,7 +686,7 @@ pub async fn all_data_types_roundtrip(db_client: &DatabaseClient) -> anyhow::Res
     let read_tx = db_client.single_use().build();
     let mut result_set = read_tx
         .execute_read(
-            google_cloud_spanner::ReadRequest::builder(
+            ReadRequest::builder(
                 "AllTypes",
                 vec![
                     "ColBool",
@@ -710,7 +714,7 @@ pub async fn all_data_types_roundtrip(db_client: &DatabaseClient) -> anyhow::Res
                 ],
             )
             .with_keys(
-                google_cloud_spanner::KeySet::builder()
+                KeySet::builder()
                     .add_key(google_cloud_spanner::key![id])
                     .build(),
             )
@@ -868,7 +872,7 @@ pub async fn all_data_types_roundtrip(db_client: &DatabaseClient) -> anyhow::Res
     let read_tx_null = db_client.single_use().build();
     let mut result_set_null = read_tx_null
         .execute_read(
-            google_cloud_spanner::ReadRequest::builder(
+            ReadRequest::builder(
                 "AllTypes",
                 vec![
                     "ColBool",
@@ -896,7 +900,7 @@ pub async fn all_data_types_roundtrip(db_client: &DatabaseClient) -> anyhow::Res
                 ],
             )
             .with_keys(
-                google_cloud_spanner::KeySet::builder()
+                KeySet::builder()
                     .add_key(google_cloud_spanner::key![id_null])
                     .build(),
             )
